Update type data Metric value to Double in model ResponseLocationWeather.kt file

+++ app/src/main/java/com/ysn/cuacain/model/weather/locationweather/ResponseLocationWeather.kt
@@ -1,8 +1,8 @@
 /*
- * Created by YSN Studio on 5/20/18 1:58 PM
+ * Created by YSN Studio on 5/23/18 4:46 AM
  * Copyright (c) 2018. All rights reserved.
  *
- * Last modified 5/20/18 1:49 PM
+ * Last modified 5/23/18 4:43 AM
  */
 
 package com.ysn.cuacain.model.weather.locationweather
@@ -57,7 +57,7 @@ data class Elevation(
 )
 
 data class Metric(
-    @SerializedName("Value") val value: Int?,
+    @SerializedName("Value") val value: Double?,
     @SerializedName("Unit") val unit: String?,
     @SerializedName("UnitType") val unitType: Int?
 )
